Draw vectors head to tail
To do this we need to calculate all the partial sums. Since I don't have
internet access, I implemented my own vector sum function as well as
something that looks like `scanl`.

Of note, the Elm functions `Basics.fromPolar` and `Basics.toPolar`
expect and return values in radians, not degrees. Since we display and
accept input from the user in degrees, we need to convert back and
forth. I needed to implement a custom conversion from degrees to radians
for this. I considered storing theta in degrees or even storing the
point in cartesian coordinates but I would still need to convert to
polar (in degrees) to show to the user so decided not to change the
internal representation at this time.