Animate rotation
Because all the vectors rotate at the same speed, the effect is the
whole drawing rotating together without moving at the joints.

I need to investigate further to get the effect I need. It's possible I
need the angles of the vectors to be relative to the previous vector
rather than the axes.