Introduce frequency
All vectors have a frequency which acts as a multiplier on the base
speed (later this should become the speed itself).

This also introduces separate concepts of a `Vector` and `Polar`.
`Vector` is the object being manipulated by the system. `Polar` is just
a normal point in some coordinate space that can be used to draw.